backend/python-service/requirements.txt
telethon==1.37.0
aiohttp==3.9.1
python-dotenv==1.0.0
cryptg==0.4.0
pysocks==1.7.1
opentele==1.15.1



